+++ airbyte-cdk/java/airbyte-cdk/s3-destinations/src/main/kotlin/io/airbyte/cdk/integrations/destination/s3/SerializedBufferFactory.kt
@@ -45,7 +45,8 @@ class SerializedBufferFactory {
         @JvmStatic
         fun getCreateFunction(
             config: S3DestinationConfig,
-            createStorageFunctionWithoutExtension: Function<String, BufferStorage>
+            createStorageFunctionWithoutExtension: Function<String, BufferStorage>,
+            useV2FieldNames: Boolean = false,
         ): BufferCreateFunction {
             val formatConfig = config.formatConfig!!
             logger.info { "S3 format config: $formatConfig" }
@@ -58,7 +59,8 @@ class SerializedBufferFactory {
                     }
                     return AvroSerializedBuffer.createFunction(
                         formatConfig as UploadAvroFormatConfig,
-                        createStorageFunctionWithExtension
+                        createStorageFunctionWithExtension,
+                        useV2FieldNames
                     )
                 }
                 FileUploadFormat.CSV -> {
@@ -69,7 +71,8 @@ class SerializedBufferFactory {
                     }
                     return CsvSerializedBuffer.createFunction(
                         formatConfig as UploadCsvFormatConfig,
-                        createStorageFunctionWithExtension
+                        createStorageFunctionWithExtension,
+                        useV2FieldNames
                     )
                 }
                 FileUploadFormat.JSONL -> {
@@ -81,13 +84,14 @@ class SerializedBufferFactory {
                     return JsonLSerializedBuffer.createBufferFunction(
                         formatConfig as UploadJsonlFormatConfig,
                         createStorageFunctionWithExtension,
+                        useV2FieldNames
                     )
                 }
                 FileUploadFormat.PARQUET -> {
                     // we can't choose the type of buffer storage with parquet because of how the
                     // underlying hadoop
                     // library is imposing file usage.
-                    return ParquetSerializedBuffer.createFunction(config)
+                    return ParquetSerializedBuffer.createFunction(config, useV2FieldNames)
                 }
                 else -> {
                     throw RuntimeException("Unexpected output format: ${Jsons.serialize(config)}")